add tps65185 command for the TI TPS65185 HVPMIC
Add tps65185 command to use all the API provided by liblphwtools
and control the TI TPS65185 HVPMIC.  This enables the power
sequences, VCOM register and HV output status to be controlled
manually from user-space.

Note: This is not yet included in the `power' command.  On some
hardware, the VCOM switch can only be controlled via a GPIO which
is not yet supported here.

+ tweak error reporting and help message wording
+ use <plsdk/log.h>