eeprom: enable EEPROM I2C address from config file
When `-o addr=CONFIG' is used, the EEPROM I2C address is searched
in the config file under the [plhwtools] section with the key
specified by CONFIG.  If not found, the value passed with -a will
be used.  If -a value is provided then plhwtools will abort and
report an error.